refactor: re-organized ArtistScreen.kt.
